cant delete user that register to tours or guide tours

+++ Controllers/UserController.cs
@@ -199,6 +199,23 @@ namespace yyytours.Controllers
             if (user == null)
                 return View("NotFound");
 
+            bool isGuideTours = _context.Trip.Count(t => t.GuideId == email) > 0;
+            // todo: when assa finish registration, add 
+            //  isSignToTours = _context.TripRegistration.Count(r => r.userEmail == email) > 0;
+            bool isSignToTours = false;
+
+            if (isGuideTours)
+            {
+                ModelState.AddModelError("userGuideTours", "לא ניתן למחוק מדריך שמדריך טיולים עתידיים");
+                return View(user);
+            }
+
+            if (isSignToTours)
+            {
+                ModelState.AddModelError("userSignToTours", "לא ניתן למחוק משתמש שרשום לטיולים");
+                return View(user);
+            }
+
             _context.User.Remove(user);
             await _context.SaveChangesAsync();
